[Refactor] AppShortInfoSwipeView 하위 뷰로 분리
- 앱 정보 수평 스크롤 영역 하위 뷰로 분리
