+++ 35-Seminar/Week2/Views/AppTitleInfo/AppTitleInfoView.swift
@@ -6,6 +6,8 @@
 //
 
 import UIKit
+import SnapKit
+import Then
 
 class AppTitleInfoView: UIView {
